Change properties.xml file before deploying.
The IRProject makes use of WordNet by using JWNL (Java WordNet Library) which expects a properties.xml file to point to the actual absolute path to the WordNet directory. The properties.xml file in IRProject.war file needs to be edited, and the value attribute of the following element needs to point to the actual WordNet directory which may be different based on Tomcat directory.

<param name="dictionary_path" value=”specify wordnet directory path here, it should end with …tomcat\webapps\irproject\wordnet\dict”/>

This path is system dependent and will be different for Linux. This properties file can be edited before deploying it to Tomcat by editing it in the war file, in the packaged war file the properties.xml is located at /WEB-INF/classes/properties.xml. The war file is just an archive file, it could be opened and edited by 7zip program.